be independent of user language
On my system the latter part of the error message is internationalized,
but the former is currently hard-coded and is therefore safe to use for
now.

Better solutions:
a) Force English language for Windows error messages. I do not know if
   this is possible.
b) Generate internationalized error message from GetLastError, same as
   libgit2 internal code.